\name{gmlView}
\alias{gmlView}
\title{Open graph in gml format}
\usage{
  gmlView()
}
\description{
  This function is used to visualize relation between
  individual sequence clusters. Graph derived from groups
  of mutually connected clusters is loaded from .gml file.
  Layout is caclulated using Frucherman-Reingold algorithm
  and shown in interactive plot. Size of vertices is
  proportional tu cluster size. Width of edges is
  proportional to numbe of mutually similar reads Function
  does not take any parameters.
}
\examples{
## do nut run!
## gmlView()
}

