% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqGrapheR-package.R
\docType{package}
\name{SeqGrapheR-package}
\alias{SeqGrapheR-package}
\title{Simple GUI for graph based visualization of cluster of DNA sequence reads}
\description{
Simple GUI for graph based visualization of cluster of DNA sequence reads
}
\details{
SeqGrapheR package provide interactive GUI for visualization of DNA sequence clusters.
Detailes and principles of usage are described in user manual and in BMC Bioinformatics 11: 378.For full
functionality installed NCBI blast is required
}
\author{
Petr Novak \email{petr<at>umbr.cas.cz}
}
\keyword{package}
