% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqGrapheR.R
\name{SeqGrapheR}
\alias{SeqGrapheR}
\title{SeqGrapheR - graph based interactive visualization of DNA sequence clusters}
\usage{
SeqGrapheR()
}
\description{
Function start GUI for interactive graph based visualization of DNA sequence clusters.
The GUI is written using the gWidgets and use RGtk2 graphical toolkit. This function
does not take any parameters. Note that SeqGrapheR require external programs
blastall and dotter for some functions.
}
\examples{
## do nut run!
## SeqGrapheR()
}
